reorganize project settings
