+++ apps/builder/app/builder/builder.tsx
@@ -48,7 +48,7 @@ import { useSyncPageUrl } from "~/shared/pages";
 import { useMount } from "~/shared/hook-utils/use-mount";
 import { subscribeCommands } from "~/builder/shared/commands";
 import { AiCommandBar } from "./features/ai/ai-command-bar";
-import { ProjectSettings } from "./features/seo/project-settings";
+import { ProjectSettings } from "./features/project-settings";
 import type { UserPlanFeatures } from "~/shared/db/user-plan-features.server";
 import { $userPlanFeatures } from "./shared/nano-states";
 import { useNavigatorLayout } from "./features/sidebar-left/navigator";
