+++ apps/builder/app/builder/features/project-settings/index.ts
@@ -1,0 +1,2 @@
+export * from "./project-settings";
+export * from "./image-control";
